pyproject.toml
[project]
name = "familien-dashboard-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "gunicorn>=23.0.0",
    "requests>=2.32.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes (unused imports, undefined names)
    "I",   # isort
    "UP",  # pyupgrade
]

[tool.ruff.lint.per-file-ignores]
"migrations/*" = ["E501", "W291"]
"tests/*" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
